+++ include/gui/QConditionalControl.h
@@ -8,17 +8,44 @@
 #include <QCheckBox>
 #include <QHBoxLayout>
 
+
+/**
+ * Виджет обертка, добавлющий к обарчиваемому виджету
+ * checkbox активности оборачиваемого виджета.
+ * Если checkbox, то и обарачивемый виджет активен.
+ */
 class QConditionalControl : public QWidget {
 
 public:
 
     QConditionalControl();
+    /**
+     * Добавить компоновщик
+     * @param layout  компоновка
+     * @param stretch
+     */
     void addLayout(QLayout *layout, size_t stretch = 0);
+    /**
+     * Добавить виджет
+     * @param widget виджет
+     * @param stretch
+     */
     void addWidget(QWidget *widget, size_t stretch = 0);
+    /**
+     * Проверить активность виджета
+     * @return true - виджет активен (галочка установлена)
+     */
     bool isChecked();
+    /**
+     * Установка режима работы
+     * @param disableChangingState true - перевести в отключенный режим (сянть галочку)
+     * false - перевести в активный режим (установить галочку)
+     */
     void setDisableChangingState(bool disableChangingState);
+    /// Проверка на отключенность виджета
     bool isDisableChangingState();
 private slots:
+    /// Слот для обновления состояния виджетов
     void stateUpdate(int state);
 private:
     struct Controller  {
