+++ include/gui/QISddStateWidget.h
@@ -10,11 +10,27 @@
 #include <QWidget>
 #include <memory>
 
+/**
+ * Интерфейс виджета получения доступа к состоянию Сферчиеского привода
+ */
 class QISddStateWidget : public QWidget {
 public:
+    /// Получить последние доступные состояния устройства
     virtual std::vector<sdd::conn::State> getSddStates() = 0;
+    /**
+     * Загрузить настройки
+     * @param settings объект доступа к настройкам
+     */
     virtual void settingsLoad(const QSettings &settings) = 0;
+    /**
+     * Сохранить настройки
+     * @param settings объект доступа к настройкам
+     */
     virtual void settingsStore(QSettings &settings) = 0;
+    /**
+     * Получить реализацию интерфейса доступа к устройству
+     * @return указатель на объект, реализующий доступ
+     */
     virtual std::shared_ptr<sdd::conn::QIConnection> getSddConnection() = 0;
     // addCallbackStateReceived()
     // removeCallbackStateReceived()
