+++ include/model/QSddModelExecutor.h
@@ -12,8 +12,9 @@
 #include <thread>
 
 /**
+ * Адаптер.
  * Класс выполняющий организацию связи модели sdd с системой сообщений QT,
- * а так же запуск моделирования в отдельном пототке.
+ * а так же запуск моделирования в отдельном потоке.
  */
 class QSddModelExecutor : public sdd::conn::QIConnection {
     Q_OBJECT
