+++ include/model/SddModel.h
@@ -23,39 +23,59 @@ public:
      * Состояние моделируемого устройства
      */
     struct State {
-        double positionOx; /// Радианы поворота устройства по оси OX
-        double positionOz;/// Радианы поворота устройства по оси OZ
-        double speedOx; /// Скорость в радианах в секунду по оси OX
-        double speedOz; /// Скорость в радианах в секунду по оси OZ
-        double IOx; /// Сила тока на горизонатльных парных соосных катушках
-        double IOz; /// Сила тока на вертикальных парных соосных катушках
-        double time; /// момент времени моделирования, в котором модель имела данно состояние
-        double oxSignal; // Воходной сигнал в модель
-        double ozSignal; // Входной сигнал в модель
+        /// Радианы поворота устройства по оси OX
+        double positionOx;
+        /// Радианы поворота устройства по оси OZ
+        double positionOz;
+        /// Скорость в радианах в секунду по оси OX
+        double speedOx;
+        /// Скорость в радианах в секунду по оси OZ
+        double speedOz;
+        /// Сила тока на горизонатльных парных соосных катушках
+        double IOx;
+        /// Сила тока на вертикальных парных соосных катушках
+        double IOz;
+        /// момент времени моделирования, в котором модель имела данно состояние
+        double time;
+        /// Воходной сигнал в модель
+        double oxSignal;
+        /// Входной сигнал в модель
+        double ozSignal;
     };
     /**
      * Вход в модель. Может изменятся во время моделирования. Описывает воходное напряжение на катушки
      * Может использоваться для внешного моделирования шим контроллера, меняя напряжение в модели
      */
     struct Input {
-        double pwmOx; /// Напряжение  на горизонатльных парных соосных катушках
-        double pwmOz;/// Напряжение  на вертикальных  парных соосных катушках
+        /// Напряжение  на горизонатльных парных соосных катушках
+        double pwmOx;
+        /// Напряжение  на вертикальных  парных соосных катушках
+        double pwmOz;
         bool operator==(const Input &other) const {
             return (this->pwmOx == other.pwmOx) && (this->pwmOz == other.pwmOz);
         }
     };
     /**
-     * Параметры модели. Могут быть изменены перед инициальзацией модели. \see init()
+     * Параметры модели. Могут быть изменены перед инициальзацией модели.
+     * @see init()
      */
     struct Parameters {
-        double positionOx0; /// Начальное положение по оси OХ в радианах
-        double positionOz0; /// Начальное положение по оси OZ в радианах
-        double speedOx0; /// Начальная скорость по оси OX в радианах в секунду
-        double speedOz0; /// Начальная скорость по оси OZ в радианах в секунду
-        double structCoeff; /// Коэффициент модели описывющй геометрию модели. Является обратным значением к моменту инерции
-        double frictionCoeff;/// Коэффициент составлющий момента трения
-        double frictionLinearCoeff;/// Момент трения состоит из двух частей, линейнозависящей от скорости и квадратично зависящей от скорости. Это коэффициент линейной части
-        double frictionQuadraticCoeff; /// Момент трения состоит из двух частей, линейнозависящей от скорости и квадратично зависящей от скорости. Это коэффициент квадратичной части
+        /// Начальное положение по оси OХ в радианах
+        double positionOx0;
+        /// Начальное положение по оси OZ в радианах
+        double positionOz0;
+        /// Начальная скорость по оси OX в радианах в секунду
+        double speedOx0;
+        /// Начальная скорость по оси OZ в радианах в секунду
+        double speedOz0;
+        /// Коэффициент модели описывющй геометрию модели. Является обратным значением к моменту инерции
+        double structCoeff;
+        /// Коэффициент составлющий момента трения
+        double frictionCoeff;
+        /// Момент трения состоит из двух частей, линейнозависящей от скорости и квадратично зависящей от скорости. Это коэффициент линейной части
+        double frictionLinearCoeff;
+        /// Момент трения состоит из двух частей, линейнозависящей от скорости и квадратично зависящей от скорости. Это коэффициент квадратичной части
+        double frictionQuadraticCoeff;
     };
 
     SddModel();
@@ -71,17 +91,19 @@ public:
      */
     bool isInit();
     /*
-     * Сброс сотояния модели. Для запуска моделирования необходимо инициализация \see init()
+     * Сброс сотояния модели. Для запуска моделирования необходимо инициализация
+     * @see init()
      */
     void reset();
     /**
-     * Выполнение одного шага моделирования. Перед вызовом этой функции необходимо выполнить инициализацию модели \see init()
+     * Выполнение одного шага моделирования. Перед вызовом этой функции необходимо выполнить инициализацию модели
+     * @see init()
      * @return Состояние модели на текущем шаге моделирования
      */
     State step();
     /**
-     * Установка входиных значений в модель. ПО умолчанию значение воходных параметров раны нулю
-     * @param input - структура вхожных параметров
+     * Установка входиных значений в модель. По умолчанию значение воходных параметров равны нулю
+     * @param input - структура входных параметров
      */
     void setInput(const Input &input);
     /**
@@ -90,10 +112,11 @@ public:
      */
     Parameters getParameters();
     /**
-     * Установка новых параметров модели. Выполняется перед инициализацией \see init().
+     * Установка новых параметров модели. Выполняется перед инициализацией.
      * Параметры не могут менятся во время моделирования. Если необходимо изменить парметры уже инициализированной модели,
-     * необходимо выполнить сброс модели \see reset(), установить праметры, а затем проинициализировать модель \see init().
+     * необходимо выполнить сброс модели (reset()), установить праметры, а затем проинициализировать модель (inti()).
      * @param parameters - параметры модели
+     * @see init(); reset()
      */
     void setParameters(const Parameters &parameters);
     /**
