+++ src/camera/JpgBuffConverter.cpp
@@ -30,13 +30,13 @@ public:
         cinfo.err = jpeg_std_error(&jerr.pub);
         jerr.pub.error_exit = my_error_exit;
         /* Establish the setjmp return context for my_error_exit to use. */
-        if (setjmp(jerr.setjmp_buffer)) {
-            /* If we get here, the JPEG code has signaled an error.
-             * We need to clean up the JPEG object, close the input file, and return.
-             */
-            jpeg_destroy_decompress(&cinfo);
-            return;
-        }
+//        if (setjmp(jerr.setjmp_buffer)) {
+//            /* If we get here, the JPEG code has signaled an error.
+//             * We need to clean up the JPEG object, close the input file, and return.
+//             */
+//            jpeg_destroy_decompress(&cinfo);
+//            return;
+//        }
         /* Now we can initialize the JPEG decompression object. */
         jpeg_create_decompress(&cinfo);
     }
